updated code according to Andrea's comments
